+++ click_and_drop_api/simple/shipping_options/large_letter.py
@@ -1,0 +1,12 @@
+"""Shipping options for large letters."""
+
+from pathlib import Path
+
+from .parser import parse_file
+
+_dir = Path(__file__).parent
+
+options = [
+    *parse_file(_dir / "large_letter-GB.txt", "largeLetter"),
+    *parse_file(_dir / "large_letter-DE.txt", "largeLetter"),
+]