+++ click_and_drop_api/simple/shipping_options/letter.py
@@ -1,0 +1,12 @@
+"""Shipping options for letters."""
+
+from pathlib import Path
+
+from .parser import parse_file
+
+_dir = Path(__file__).parent
+
+options = [
+    *parse_file(_dir / "letter-GB.txt", "letter"),
+    *parse_file(_dir / "letter-DE.txt", "letter"),
+]